+++ apps/zhi-cli/CHANGELOG.md
@@ -1,5 +1,21 @@
 # zhi-cli
 
+## 1.5.6
+
+### Patch Changes
+
+- add hash query string
+- Updated dependencies
+  - zhi-log@1.14.6
+
+## 1.5.5
+
+### Patch Changes
+
+- fix esbuild deps
+- Updated dependencies
+  - zhi-log@1.14.5
+
 ## 1.5.4
 
 ### Patch Changes