+++ libs/zhi-blog-api/CHANGELOG.md
@@ -1,5 +1,21 @@
 # zhi-blog-api
 
+## 1.6.6
+
+### Patch Changes
+
+- add hash query string
+- Updated dependencies
+  - zhi-lib-base@0.2.6
+
+## 1.6.5
+
+### Patch Changes
+
+- fix esbuild deps
+- Updated dependencies
+  - zhi-lib-base@0.2.5
+
 ## 1.6.4
 
 ### Patch Changes