+++ libs/zhi-common/CHANGELOG.md
@@ -1,5 +1,21 @@
 # zhi-common
 
+## 1.9.6
+
+### Patch Changes
+
+- add hash query string
+- Updated dependencies
+  - zhi-lib-base@0.2.6
+
+## 1.9.5
+
+### Patch Changes
+
+- fix esbuild deps
+- Updated dependencies
+  - zhi-lib-base@0.2.5
+
 ## 1.9.4
 
 ### Patch Changes