+++ libs/zhi-env/CHANGELOG.md
@@ -1,5 +1,17 @@
 # zhi-env
 
+## 1.18.6
+
+### Patch Changes
+
+- add hash query string
+
+## 1.18.5
+
+### Patch Changes
+
+- fix esbuild deps
+
 ## 1.18.4
 
 ### Patch Changes