+++ tools/commit-config-custom/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @terwer/commit-config-custom
 
+## 1.0.9
+
+### Patch Changes
+
+- add hash query string
+
+## 1.0.8
+
+### Patch Changes
+
+- fix esbuild deps
+
 ## 1.0.7
 
 ### Patch Changes