+++ tools/tsconfig/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @terwer/tsconfig
 
+## 0.0.7
+
+### Patch Changes
+
+- add hash query string
+
+## 0.0.6
+
+### Patch Changes
+
+- fix esbuild deps
+
 ## 0.0.5
 
 ### Patch Changes