+++ tools/vite-config-custom/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @terwer/vite-config-custom
 
+## 0.7.6
+
+### Patch Changes
+
+- add hash query string
+
+## 0.7.5
+
+### Patch Changes
+
+- fix esbuild deps
+
 ## 0.7.4
 
 ### Patch Changes