+++ tools/vitest-config-custom/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @terwer/vitest-config-custom
 
+## 1.0.6
+
+### Patch Changes
+
+- add hash query string
+
+## 1.0.5
+
+### Patch Changes
+
+- fix esbuild deps
+
 ## 1.0.4
 
 ### Patch Changes